+++ packages/dart/sshnp_flutter/lib/l10n/app_en.arb
@@ -43,16 +43,12 @@
     "from" : "From",
     "homeDirectory" : "Home Directory",
     "homeDirectoryHint" : "The home directory on this host",
-    "host" : "SR Address (atSign)",
-    "hostTooltip" : "atSign of srvd daemon or FQDN/IP address to connect back to",
-    "hostSelection" : "Host Selection",
     "import" : "Import",
     "importProfile" : "Import Profile",
     "keyFile" : "Key File",
     "listDevices" : "List Devices",
-    "localPort": "Local Forwarding Port",
-    "localPortTooltip" : "Port on which sshd is listening locally on the client host",
-    "localSshdPort": "Local SSHD Port",
+    "localPort": "Client-side port for the ssh tunnel",
+    "localPortTooltip" : "Client-side port for the ssh tunnel. Defaults to 0 (ask the o/s for a port)",
     "localSshOptions" : "Local SSH Options",
     "localSshOptionsTooltip" : "Add these commands to the local ssh command",
     "localSshOptionsHint" : "Use \",\" to separate options",
@@ -65,7 +61,6 @@
     "noTerminalSessionsHelp" : "Create a new session from the home screen",
     "okButton" : "Ok",
     "options" : "Options",
-    "port" : "Remote Port",
     "privacyPolicy" : "Privacy Policy",
     "privateKey" : "SSH Private Key",
     "privateKeyTooltip" : "Private Key for authentication",
@@ -80,7 +75,7 @@
     "profileName" : "Profile Name",
     "profileNameTooltip" : "Name of the profile to use",
     "remoteSshdPort": "Remote SSHD Port",
-    "remoteSshdPortTooltip" : "Port on which sshd is listening locally on the device host",
+    "remoteSshdPortTooltip" : "The device's sshd's listening port",
     "remoteUserName" : "Session username",
     "remoteUserNameTooltip" : "Username to use in the ssh session on the remote host",
     "removeButton":"Remove",
@@ -103,6 +98,8 @@
     "supportDescription" : "Our team of experts is here to help! Select your preferred method below",
     "sourcePort" : "Source Port",
     "socketRendezvousConfiguration" : "Socket Rendezvous\nConfiguration",
+    "srvdAtsign" : "SR Address (atSign)",
+    "srvdAtsignTooltip" : "atSign of the socket rendezvous",
     "sshAlgorithm" : "SSH Algorithm",
     "sshButton" : "ssh",
     "sshKeyManagement": "{newLine, select, yes{SSH Key \nConfiguration} no{SSH Key Management} other{SSH Key Management}}",
