+++ packages/dart/sshnp_flutter/lib/main.dart
@@ -4,7 +4,7 @@ import 'package:at_app_flutter/at_app_flutter.dart' show AtEnv;
 import 'package:at_utils/at_logger.dart' show AtSignLogger;
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:sshnp_flutter/src/utility/platform_utility/platform_utililty.dart';
+import 'package:sshnp_flutter/src/utility/platform_utility/platform_utility.dart';
 
 import '../src/repository/authentication_repository.dart';
