+++ packages/dart/sshnp_flutter/lib/src/presentation/widgets/settings_actions/settings_reset_app_action.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:sshnp_flutter/src/presentation/widgets/settings_screen_widgets/settings_actions/settings_action_button.dart';
 import 'package:sshnp_flutter/src/presentation/widgets/utility/at_error_dialog.dart';
-import 'package:sshnp_flutter/src/utility/platform_utility/platform_utililty.dart';
+import 'package:sshnp_flutter/src/utility/platform_utility/platform_utility.dart';
 import 'package:sshnp_flutter/src/utility/sizes.dart';
 
 /// Custom reset button widget is to reset an atsign from keychain list,
