+++ packages/dart/sshnp_flutter/lib/src/presentation/widgets/terminal_screen_widgets/terminal_screen_mobile_view.dart
@@ -21,7 +21,7 @@ class _TerminalScreenMobileState extends ConsumerState<TerminalScreenMobileView>
     super.initState();
     final sessionId = ref.read(terminalSessionController);
 
-    final sessionController = ref.read(terminalSessionFamilyController(sessionId).notifier);
+    ref.read(terminalSessionFamilyController(sessionId).notifier);
     WidgetsBinding.instance.endOfFrame.then((value) {});
   }
 
@@ -43,7 +43,7 @@ class _TerminalScreenMobileState extends ConsumerState<TerminalScreenMobileView>
     final terminalList = ref.watch(terminalSessionListController);
     final currentSessionId = ref.watch(terminalSessionController);
     final int currentIndex = (terminalList.isEmpty) ? 0 : terminalList.indexOf(currentSessionId);
-    final tabController = TabController(initialIndex: currentIndex, length: terminalList.length, vsync: this);
+    TabController(initialIndex: currentIndex, length: terminalList.length, vsync: this);
 
     return Scaffold(
       appBar: CustomAppBar(
