+++ packages/dart/sshnp_flutter/lib/src/utility/platform_utility/default_platform_utility.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:sshnp_flutter/src/controllers/navigation_controller.dart';
 import 'package:sshnp_flutter/src/utility/app_theme.dart';
-import 'package:sshnp_flutter/src/utility/platform_utility/platform_utililty.dart';
+import 'package:sshnp_flutter/src/utility/platform_utility/platform_utility.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class DefaultPlatformUtility implements PlatformUtility {
