+++ packages/dart/sshnp_flutter/lib/src/utility/platform_utility/macos_utility.dart
@@ -8,7 +8,7 @@ import 'package:macos_ui/macos_ui.dart';
 import 'package:sshnp_flutter/src/controllers/navigation_controller.dart';
 import 'package:sshnp_flutter/src/utility/app_theme.dart';
 import 'package:sshnp_flutter/src/utility/platform_utility/default_platform_utility.dart';
-import 'package:sshnp_flutter/src/utility/platform_utility/platform_utililty.dart';
+import 'package:sshnp_flutter/src/utility/platform_utility/platform_utility.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class MacosUtility implements PlatformUtility {
@@ -17,9 +17,6 @@ class MacosUtility implements PlatformUtility {
   @override
   Future<void> configurePlatform() async {
     return;
-    const config =
-        MacosWindowUtilsConfig(toolbarStyle: NSWindowToolbarStyle.unified);
-    await config.apply();
   }
 
   @override
