+++ packages/dart/sshnp_flutter/lib/src/utility/platform_utility/platform_utility.dart
