+++ packages/dart/sshnp_flutter/pubspec.yaml
@@ -47,7 +47,8 @@ dependencies:
   intl: any
   macos_ui: ^2.0.0
   mocktail: ^1.0.0
-  noports_core: 6.0.1
+  noports_core:
+    path: ../noports_core/
   page_transition: ^2.0.9
   path: ^1.8.3
   path_provider: ^2.1.1
